refactored startEvaluationRunPage
