+++ packages/eval-hub/frontend/src/app/pages/StartEvaluationRunPage.tsx
@@ -32,345 +32,47 @@ import ApplicationsPage from '@odh-dashboard/internal/pages/ApplicationsPage';
 import {
   MlflowExperimentSelector,
   useMlflowExperiments,
-  type MlflowExperiment,
 } from '@odh-dashboard/internal/concepts/mlflow';
-import {
-  fireFormTrackingEvent,
-  fireMiscTrackingEvent,
-} from '@odh-dashboard/internal/concepts/analyticsTracking/segmentIOUtils';
-import { TrackingOutcome } from '@odh-dashboard/internal/concepts/analyticsTracking/trackingProperties';
-import { createEvaluationJob } from '~/app/api/k8s';
-import { EVAL_HUB_EVENTS } from '~/app/tracking/evalhubTrackingConstants';
-import buildEvaluationRequest from '~/app/utils/buildEvaluationRequest';
-import getErrorTitle from '~/app/utils/getErrorTitle';
 import {
   evaluationsBaseRoute,
   evaluationBenchmarksRoute,
   evaluationCollectionsRoute,
   evaluationCreateRoute,
 } from '~/app/routes';
-import { useNotification } from '~/app/hooks/useNotification';
 import { useEvaluationSelection } from '~/app/hooks/useEvaluationSelection';
 import LabelHelpPopover from '~/app/components/LabelHelpPopover';
 import BenchmarkThresholdField from '~/app/components/BenchmarkThresholdField';
 import PrimaryScorerMetricField from '~/app/components/PrimaryScorerMetricField';
+import {
+  useStartEvaluationRunForm,
+  EXPERIMENT_FILTER,
+  DEFAULT_EXPERIMENT_NAME,
+} from './useStartEvaluationRunForm';
 
 import './StartEvaluationRunPage.css';
-
-type InputMode = 'inference' | 'prerecorded';
-type ExperimentMode = 'existing' | 'new';
-
-const DEFAULT_EXPERIMENT_NAME = 'EvalHub';
-const EXPERIMENT_FILTER = "tags.context = 'eval-hub'";
 
 const StartEvaluationRunPage: React.FC = () => {
   const { namespace } = useParams<{ namespace: string }>();
   const navigate = useNavigate();
-  const notification = useNotification();
 
   const { benchmark, collection, isCollectionFlow, dataLoaded, loadError } =
     useEvaluationSelection(namespace);
-
-  const DEFAULT_SUITE_THRESHOLD = 70;
-
-  const defaultThreshold = React.useMemo(() => {
-    if (collection?.pass_criteria) {
-      return Math.round(collection.pass_criteria.threshold * 100);
-    }
-    if (collection) {
-      return DEFAULT_SUITE_THRESHOLD;
-    }
-    if (benchmark?.pass_criteria) {
-      return Math.round(benchmark.pass_criteria.threshold * 100);
-    }
-    return 0;
-  }, [benchmark, collection]);
-
-  const availableMetrics = React.useMemo(() => benchmark?.metrics ?? [], [benchmark]);
-
-  const defaultPrimaryMetric = benchmark?.primary_score?.metric ?? availableMetrics[0];
-
-  const [threshold, setThreshold] = React.useState(defaultThreshold);
-  const [thresholdTouched, setThresholdTouched] = React.useState(false);
-  const [primaryMetric, setPrimaryMetric] = React.useState<string | undefined>(
-    defaultPrimaryMetric,
-  );
-
-  const handleThresholdChange = React.useCallback((value: number) => {
-    setThreshold(value);
-    setThresholdTouched(true);
-  }, []);
-
-  const handlePrimaryMetricChange = React.useCallback((metric: string) => {
-    setPrimaryMetric(metric);
-  }, []);
 
   const { data: experiments, loaded: experimentsLoaded } = useMlflowExperiments({
     workspace: namespace ?? '',
     filter: EXPERIMENT_FILTER,
   });
 
-  const [evaluationName, setEvaluationName] = React.useState(() =>
-    new Date().toLocaleString('en-US', {
-      month: 'short',
-      day: 'numeric',
-      year: 'numeric',
-      hour: 'numeric',
-      minute: '2-digit',
-      hour12: true,
-    }),
-  );
-
-  const [inputMode, setInputMode] = React.useState<InputMode>('inference');
-
-  const [modelName, setModelName] = React.useState('');
-  const [endpointUrl, setEndpointUrl] = React.useState('');
-  const [apiKeySecretRef, setApiKeySecretRef] = React.useState('');
-
-  const [sourceName, setSourceName] = React.useState('');
-  const [datasetUrl, setDatasetUrl] = React.useState('');
-  const [accessToken, setAccessToken] = React.useState('');
-
-  const [experimentMode, setExperimentMode] = React.useState<ExperimentMode>('existing');
-  const [selectedExperiment, setSelectedExperiment] = React.useState<MlflowExperiment | undefined>(
-    undefined,
-  );
-  const [newExperimentName, setNewExperimentName] = React.useState('');
-  const [experimentAutoSelected, setExperimentAutoSelected] = React.useState(false);
-
-  React.useEffect(() => {
-    if (!experimentsLoaded || !namespace || experimentAutoSelected) {
-      return;
-    }
-    setExperimentAutoSelected(true);
-
-    if (experiments.length === 0) {
-      setExperimentMode('new');
-      setNewExperimentName(DEFAULT_EXPERIMENT_NAME);
-    } else {
-      const defaultExp = experiments.find((e) => e.name === DEFAULT_EXPERIMENT_NAME);
-      setExperimentMode('existing');
-      setSelectedExperiment(defaultExp ?? experiments[0]);
-    }
-  }, [experimentsLoaded, experiments, namespace, experimentAutoSelected]);
-
-  const [showAdditionalArgs, setShowAdditionalArgs] = React.useState(false);
-  const [additionalArgs, setAdditionalArgs] = React.useState('');
-  const [additionalArgsFilename, setAdditionalArgsFilename] = React.useState('');
-
-  const [isSubmitting, setIsSubmitting] = React.useState(false);
-  const abortControllerRef = React.useRef<AbortController | null>(null);
-  const experimentManuallyChangedRef = React.useRef(false);
-
-  React.useEffect(
-    () => () => {
-      abortControllerRef.current?.abort();
-    },
-    [],
-  );
-
-  const benchmarkDisplayName = React.useMemo(() => {
-    if (collection) {
-      return collection.name;
-    }
-    if (benchmark) {
-      return benchmark.name;
-    }
-    return '';
-  }, [benchmark, collection]);
+  const form = useStartEvaluationRunForm({
+    namespace,
+    benchmark,
+    collection,
+    isCollectionFlow,
+    experiments,
+    experimentsLoaded,
+  });
 
   const breadcrumbFlowLabel = isCollectionFlow ? 'Select benchmark suite' : 'Select benchmark';
-
-  const hasBenchmarks =
-    !!benchmark || (!!collection && !!collection.benchmarks && collection.benchmarks.length > 0);
-
-  const hasExperiment =
-    (experimentMode === 'existing' && !!selectedExperiment) ||
-    (experimentMode === 'new' && newExperimentName.trim() !== '');
-
-  const isValid = React.useMemo(() => {
-    if (evaluationName.trim() === '' || !hasBenchmarks || !hasExperiment) {
-      return false;
-    }
-    if (inputMode === 'inference') {
-      return modelName.trim() !== '' && endpointUrl.trim() !== '';
-    }
-    return sourceName.trim() !== '' && datasetUrl.trim() !== '';
-  }, [
-    evaluationName,
-    inputMode,
-    modelName,
-    endpointUrl,
-    sourceName,
-    datasetUrl,
-    hasBenchmarks,
-    hasExperiment,
-  ]);
-
-  const handleAdditionalArgsFileChange = React.useCallback(
-    (
-      _event: React.DragEvent<HTMLElement> | React.ChangeEvent<HTMLInputElement> | Event,
-      file: File,
-    ) => {
-      setAdditionalArgsFilename(file.name);
-      const reader = new FileReader();
-      reader.onload = () => {
-        const text = typeof reader.result === 'string' ? reader.result : '';
-        setAdditionalArgs(text);
-      };
-      reader.onerror = () => {
-        notification.error('File read failed', `Unable to read file "${file.name}".`);
-        setAdditionalArgsFilename('');
-      };
-      reader.readAsText(file);
-    },
-    [notification],
-  );
-
-  const handleAdditionalArgsTextChange = React.useCallback(
-    (_event: React.ChangeEvent<HTMLTextAreaElement>, value: string) => {
-      setAdditionalArgs(value);
-    },
-    [],
-  );
-
-  const handleAdditionalArgsClear = React.useCallback(() => {
-    setAdditionalArgsFilename('');
-    setAdditionalArgs('');
-  }, []);
-
-  const handleSubmit = async () => {
-    if (!isValid || isSubmitting) {
-      return;
-    }
-
-    setIsSubmitting(true);
-
-    const parsedArgs: Record<string, unknown> = {};
-    if (showAdditionalArgs && additionalArgs.trim()) {
-      try {
-        const parsed: unknown = JSON.parse(additionalArgs);
-        if (typeof parsed !== 'object' || parsed === null || Array.isArray(parsed)) {
-          notification.error(
-            'Invalid benchmark parameters',
-            'Benchmark parameters must be a JSON object (e.g. {"key": "value"}).',
-          );
-          setIsSubmitting(false);
-          return;
-        }
-        Object.assign(parsedArgs, parsed);
-      } catch {
-        notification.error(
-          'Invalid benchmark parameters',
-          'Benchmark parameters must be valid JSON.',
-        );
-        setIsSubmitting(false);
-        return;
-      }
-    }
-
-    const isNewExperiment = experimentMode === 'new';
-    const experimentName = isNewExperiment ? newExperimentName.trim() : selectedExperiment?.name;
-
-    const shouldIncludeThreshold = thresholdTouched || defaultThreshold > 0;
-    const passCriteriaOverride = shouldIncludeThreshold
-      ? { threshold: threshold / 100 }
-      : undefined;
-
-    const primaryScoreOverride = primaryMetric
-      ? {
-          metric: primaryMetric,
-          // eslint-disable-next-line camelcase
-          lower_is_better: benchmark?.primary_score?.lower_is_better ?? false,
-        }
-      : undefined;
-
-    const request = buildEvaluationRequest({
-      evaluationName,
-      inputMode,
-      benchmark,
-      collection,
-      modelName,
-      endpointUrl,
-      apiKeySecretRef,
-      sourceName,
-      datasetUrl,
-      accessToken,
-      additionalArgs: parsedArgs,
-      experimentName: experimentName || undefined,
-      experimentTags: undefined,
-      passCriteriaOverride,
-      primaryScoreOverride,
-    });
-
-    fireMiscTrackingEvent(EVAL_HUB_EVENTS.MLFLOW_EXPERIMENT_SELECTED, {
-      experimentSelection: isNewExperiment
-        ? 'new'
-        : !experimentManuallyChangedRef.current &&
-            selectedExperiment?.name === DEFAULT_EXPERIMENT_NAME
-          ? 'default'
-          : 'existing',
-      experimentName,
-    });
-
-    // Snapshot form state now so both success and error paths share the same properties.
-    const runTrackingProps = {
-      source: 'evaluations_page',
-      evaluationName: evaluationName.trim(),
-      sourceType:
-        inputMode === 'inference'
-          ? ('inference_endpoint' as const)
-          : ('pre_recorded_responses' as const),
-      modelName: inputMode === 'inference' ? modelName.trim() : undefined,
-      endpointOrigin: (() => {
-        if (inputMode !== 'inference') {
-          return undefined;
-        }
-        try {
-          return new URL(endpointUrl.trim()).origin;
-        } catch {
-          return undefined;
-        }
-      })(),
-      hasAPIKey: inputMode === 'inference' ? apiKeySecretRef.trim() !== '' : false,
-      sourceName: inputMode === 'prerecorded' ? sourceName.trim() : undefined,
-      hasDatasetURL: inputMode === 'prerecorded' ? datasetUrl.trim() !== '' : false,
-      hasAccessToken: inputMode === 'prerecorded' ? accessToken.trim() !== '' : false,
-      hasAdditionalArguments: showAdditionalArgs && additionalArgs.trim() !== '',
-      countOfAdditionalArguments: Object.keys(parsedArgs).length,
-    };
-
-    const controller = new AbortController();
-    abortControllerRef.current = controller;
-
-    try {
-      await createEvaluationJob('', namespace ?? '', request)({ signal: controller.signal });
-      fireFormTrackingEvent(EVAL_HUB_EVENTS.EVALUATION_RUN_STARTED, {
-        ...runTrackingProps,
-        outcome: TrackingOutcome.submit,
-        success: true,
-      });
-      notification.success(
-        'Evaluation started',
-        `Evaluation "${evaluationName}" has been started.`,
-      );
-      navigate(evaluationsBaseRoute(namespace));
-    } catch (e) {
-      if (controller.signal.aborted) {
-        return;
-      }
-      const message = e instanceof Error ? e.message : 'An unknown error occurred.';
-      fireFormTrackingEvent(EVAL_HUB_EVENTS.EVALUATION_RUN_STARTED, {
-        ...runTrackingProps,
-        outcome: TrackingOutcome.submit,
-        success: false,
-        errorName: e instanceof Error ? e.name : 'UnknownError',
-      });
-      notification.error(getErrorTitle(e, 'Failed to start evaluation'), message);
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
 
   if (!dataLoaded) {
     return (
@@ -446,8 +148,8 @@ const StartEvaluationRunPage: React.FC = () => {
             <TextInput
               id="evaluation-name"
               data-testid="evaluation-name-input"
-              value={evaluationName}
-              onChange={(_e, val) => setEvaluationName(val)}
+              value={form.evaluationName}
+              onChange={(_e, val) => form.setEvaluationName(val)}
               isRequired
             />
           </FormGroup>
@@ -468,15 +170,15 @@ const StartEvaluationRunPage: React.FC = () => {
               data-testid="experiment-mode-existing"
               name="experiment-mode"
               label="Select existing experiment"
-              isChecked={experimentMode === 'existing'}
+              isChecked={form.experimentMode === 'existing'}
               onChange={() => {
-                setExperimentMode('existing');
-                setNewExperimentName('');
-                experimentManuallyChangedRef.current = true;
+                form.setExperimentMode('existing');
+                form.setNewExperimentName('');
+                form.experimentManuallyChangedRef.current = true;
               }}
             />
 
-            {experimentMode === 'existing' && namespace && (
+            {form.experimentMode === 'existing' && namespace && (
               <div
                 className="eval-hub-start-evaluation-run__mlflow-selector"
                 style={{ marginTop: 'var(--pf-t--global--spacer--sm)' }}
@@ -484,10 +186,10 @@ const StartEvaluationRunPage: React.FC = () => {
                 <MlflowExperimentSelector
                   workspace={namespace}
                   filter={EXPERIMENT_FILTER}
-                  selection={selectedExperiment?.name}
+                  selection={form.selectedExperiment?.name}
                   onSelect={(exp) => {
-                    setSelectedExperiment(exp);
-                    experimentManuallyChangedRef.current = true;
+                    form.setSelectedExperiment(exp);
+                    form.experimentManuallyChangedRef.current = true;
                   }}
                 />
               </div>
@@ -499,25 +201,25 @@ const StartEvaluationRunPage: React.FC = () => {
                 data-testid="experiment-mode-new"
                 name="experiment-mode"
                 label="Create new experiment"
-                isChecked={experimentMode === 'new'}
+                isChecked={form.experimentMode === 'new'}
                 onChange={() => {
-                  setExperimentMode('new');
-                  setSelectedExperiment(undefined);
-                  experimentManuallyChangedRef.current = true;
-                  setNewExperimentName((prev) =>
+                  form.setExperimentMode('new');
+                  form.setSelectedExperiment(undefined);
+                  form.experimentManuallyChangedRef.current = true;
+                  form.setNewExperimentName((prev) =>
                     prev.trim() === '' ? DEFAULT_EXPERIMENT_NAME : prev,
                   );
                 }}
               />
             </div>
 
-            {experimentMode === 'new' && (
+            {form.experimentMode === 'new' && (
               <div style={{ marginTop: 'var(--pf-t--global--spacer--md)' }}>
                 <TextInput
                   id="new-experiment-name"
                   data-testid="new-experiment-name-input"
-                  value={newExperimentName}
-                  onChange={(_e, val) => setNewExperimentName(val)}
+                  value={form.newExperimentName}
+                  onChange={(_e, val) => form.setNewExperimentName(val)}
                   placeholder="Enter experiment name"
                 />
               </div>
@@ -535,19 +237,19 @@ const StartEvaluationRunPage: React.FC = () => {
               data-testid="input-mode-inference"
               name="input-mode"
               label="Inference endpoint"
-              isChecked={inputMode === 'inference'}
-              onChange={() => setInputMode('inference')}
+              isChecked={form.inputMode === 'inference'}
+              onChange={() => form.setInputMode('inference')}
               className="eval-hub-start-evaluation-run__source-inference-radio"
               body={
-                inputMode === 'inference' ? (
+                form.inputMode === 'inference' ? (
                   <Stack hasGutter>
                     <StackItem>
                       <FormGroup label="Model or agent name" isRequired fieldId="model-name">
                         <TextInput
                           id="model-name"
                           data-testid="model-name-input"
-                          value={modelName}
-                          onChange={(_e, val) => setModelName(val)}
+                          value={form.modelName}
+                          onChange={(_e, val) => form.setModelName(val)}
                           isRequired
                         />
                         <FormHelperText>
@@ -559,20 +261,18 @@ const StartEvaluationRunPage: React.FC = () => {
                         </FormHelperText>
                       </FormGroup>
                     </StackItem>
-
                     <StackItem>
                       <FormGroup label="Endpoint URL" isRequired fieldId="endpoint-url">
                         <TextInput
                           id="endpoint-url"
                           data-testid="endpoint-url-input"
-                          value={endpointUrl}
-                          onChange={(_e, val) => setEndpointUrl(val)}
+                          value={form.endpointUrl}
+                          onChange={(_e, val) => form.setEndpointUrl(val)}
                           placeholder="https://api.example.com/v1/model"
                           isRequired
                         />
                       </FormGroup>
                     </StackItem>
-
                     <StackItem>
                       <FormGroup
                         label="API key"
@@ -587,8 +287,8 @@ const StartEvaluationRunPage: React.FC = () => {
                         <TextInput
                           id="api-key"
                           data-testid="api-key-input"
-                          value={apiKeySecretRef}
-                          onChange={(_e, val) => setApiKeySecretRef(val)}
+                          value={form.apiKeySecretRef}
+                          onChange={(_e, val) => form.setApiKeySecretRef(val)}
                         />
                       </FormGroup>
                     </StackItem>
@@ -602,10 +302,10 @@ const StartEvaluationRunPage: React.FC = () => {
               data-testid="input-mode-prerecorded"
               name="input-mode"
               label="Pre-recorded responses"
-              isChecked={inputMode === 'prerecorded'}
-              onChange={() => setInputMode('prerecorded')}
+              isChecked={form.inputMode === 'prerecorded'}
+              onChange={() => form.setInputMode('prerecorded')}
               body={
-                inputMode === 'prerecorded' ? (
+                form.inputMode === 'prerecorded' ? (
                   <Stack hasGutter>
                     <StackItem>
                       <FormGroup
@@ -622,25 +322,23 @@ const StartEvaluationRunPage: React.FC = () => {
                         <TextInput
                           id="source-name"
                           data-testid="source-name-input"
-                          value={sourceName}
-                          onChange={(_e, val) => setSourceName(val)}
+                          value={form.sourceName}
+                          onChange={(_e, val) => form.setSourceName(val)}
                           isRequired
                         />
                       </FormGroup>
                     </StackItem>
-
                     <StackItem>
                       <FormGroup label="Dataset URL" isRequired fieldId="dataset-url">
                         <TextInput
                           id="dataset-url"
                           data-testid="dataset-url-input"
-                          value={datasetUrl}
-                          onChange={(_e, val) => setDatasetUrl(val)}
+                          value={form.datasetUrl}
+                          onChange={(_e, val) => form.setDatasetUrl(val)}
                           isRequired
                         />
                       </FormGroup>
                     </StackItem>
-
                     <StackItem>
                       <FormGroup
                         label="Access token"
@@ -655,8 +353,8 @@ const StartEvaluationRunPage: React.FC = () => {
                         <TextInput
                           id="access-token"
                           data-testid="access-token-input"
-                          value={accessToken}
-                          onChange={(_e, val) => setAccessToken(val)}
+                          value={form.accessToken}
+                          onChange={(_e, val) => form.setAccessToken(val)}
                           placeholder="e.g. my-dataset-credentials"
                         />
                       </FormGroup>
@@ -672,22 +370,22 @@ const StartEvaluationRunPage: React.FC = () => {
             fieldId="benchmark-name"
           >
             <Content component="p" data-testid="benchmark-name-display">
-              {benchmarkDisplayName}
+              {form.benchmarkDisplayName}
             </Content>
           </FormGroup>
 
           <BenchmarkThresholdField
-            value={threshold}
-            onChange={handleThresholdChange}
+            value={form.threshold}
+            onChange={form.handleThresholdChange}
             label={isCollectionFlow ? 'Benchmark suite threshold' : 'Benchmark threshold'}
             fieldId="benchmark-threshold"
           />
 
-          {!isCollectionFlow && availableMetrics.length > 0 && (
+          {!isCollectionFlow && form.availableMetrics.length > 0 && (
             <PrimaryScorerMetricField
-              metrics={availableMetrics}
-              selected={primaryMetric}
-              onChange={handlePrimaryMetricChange}
+              metrics={form.availableMetrics}
+              selected={form.primaryMetric}
+              onChange={form.handlePrimaryMetricChange}
             />
           )}
 
@@ -697,8 +395,8 @@ const StartEvaluationRunPage: React.FC = () => {
                 id="show-additional-args"
                 data-testid="show-additional-args"
                 label="Benchmark parameters"
-                isChecked={showAdditionalArgs}
-                onChange={(_e, checked) => setShowAdditionalArgs(checked)}
+                isChecked={form.showAdditionalArgs}
+                onChange={(_e, checked) => form.setShowAdditionalArgs(checked)}
               />
             </FlexItem>
             <FlexItem>
@@ -708,18 +406,18 @@ const StartEvaluationRunPage: React.FC = () => {
               />
             </FlexItem>
           </Flex>
-          {showAdditionalArgs && (
+          {form.showAdditionalArgs && (
             <FormGroup fieldId="additional-args">
               <FileUpload
                 id="additional-args"
                 data-testid="additional-args-upload"
                 type="text"
-                value={additionalArgs}
-                filename={additionalArgsFilename}
+                value={form.additionalArgs}
+                filename={form.additionalArgsFilename}
                 filenamePlaceholder="Drag and drop a file or upload"
-                onFileInputChange={handleAdditionalArgsFileChange}
-                onTextChange={handleAdditionalArgsTextChange}
-                onClearClick={handleAdditionalArgsClear}
+                onFileInputChange={form.handleAdditionalArgsFileChange}
+                onTextChange={form.handleAdditionalArgsTextChange}
+                onClearClick={form.handleAdditionalArgsClear}
                 browseButtonText="Upload"
                 allowEditingUploadedText
                 textAreaPlaceholder={'{\n  "num_examples": 10\n}'}
@@ -739,29 +437,16 @@ const StartEvaluationRunPage: React.FC = () => {
             <Button
               variant="primary"
               data-testid="start-evaluation-submit"
-              onClick={handleSubmit}
-              isDisabled={!isValid || isSubmitting}
-              isLoading={isSubmitting}
+              onClick={form.handleSubmit}
+              isDisabled={!form.isValid || form.isSubmitting}
+              isLoading={form.isSubmitting}
             >
               Start evaluation run
             </Button>
             <Button
               variant="link"
               data-testid="start-evaluation-cancel"
-              onClick={() => {
-                fireFormTrackingEvent(EVAL_HUB_EVENTS.EVALUATION_RUN_STARTED, {
-                  source: 'evaluations_page',
-                  evaluationName: evaluationName.trim(),
-                  sourceType:
-                    inputMode === 'inference'
-                      ? ('inference_endpoint' as const)
-                      : ('pre_recorded_responses' as const),
-                  hasAPIKey: inputMode === 'inference' ? apiKeySecretRef.trim() !== '' : false,
-                  hasAdditionalArguments: showAdditionalArgs && additionalArgs.trim() !== '',
-                  outcome: TrackingOutcome.cancel,
-                });
-                navigate(evaluationsBaseRoute(namespace));
-              }}
+              onClick={form.handleCancel}
             >
               Cancel
             </Button>
